Added Study Bot to Examples lists
Study Bot in C# is an example of the LUIS web app bot that integrates QnA Maker and Bing Spell Check, showing the code for the bot and these additional services, as well as using the bot inside a UWP app. This is unlike the other bot samples which don't offer other Cognitive Service integrations, so this expands the sample list capability choice for users.